app/requirements.txt
gradio==5.22.0
openai==1.70.0
requests==2.32.3
python-dotenv==1.1.0
pymongo==4.11.3
pytz==2025.2
notebook
